api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
clickhouse-connect>=0.7
asyncpg>=0.29
pyyaml>=6.0
python-dotenv>=1.0
